Break code generator into a separate subproject
